Backmerge PM-360: Logout Button
